requirements.txt
pytest
coverage
Gitpython
Pillow
requests
tkinter
tk